+++ src/tabular_trees/xgboost/__init__.py
@@ -1,5 +1,6 @@
 """Module containing functionality for dealing with xgboost models."""
 
 from . import explainer
+from . import trees
 from . import parser
 from . import validate
